fix(tui): wrap prose using display width
